+++ IsletPipeline/img_corr_fun.m
@@ -2,7 +2,7 @@ function [im_reg] = img_corr_fun(im,im_ad,frames)
 %IMG_CORR_FUN 此处显示有关此函数的摘要
 %   此处显示详细说明
 n=frames;
-im_reg = uint16(zeros(size(im_ad(:,:,1),1),size(im_ad(:,:,1),2),n));
+im_reg = single(zeros(size(im_ad(:,:,1),1),size(im_ad(:,:,1),2),n));
 tic
 im_reg(:,:,1) = im(:,:,1); 
 move_mark = zeros(n,3);
